# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SOLOS+ ESPORTZ clan management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scrims:
    get:
      operationId: listScrims
      tags: [scrims]
      summary: List all scrims
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, ongoing, completed, cancelled]
      responses:
        "200":
          description: Scrims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scrim"

    post:
      operationId: createScrim
      tags: [scrims]
      summary: Create scrim (management only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrimInput"
      responses:
        "201":
          description: Created scrim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scrim"
  /scrims/{id}:
    get:
      operationId: getScrim
      tags: [scrims]
      summary: Get scrim by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scrim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scrim"

    patch:
      operationId: updateScrim
      tags: [scrims]
      summary: Update scrim (management only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrimUpdate"
      responses:
        "200":
          description: Updated scrim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scrim"

    delete:
      operationId: deleteScrim
      tags: [scrims]
      summary: Delete scrim (management only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /scrims/{id}/join:
    post:
      operationId: joinScrim
      tags: [scrims]
      summary: Sign up for a scrim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Signup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrimSignup"
  /scrims/{id}/leave:
    post:
      operationId: leaveScrim
      tags: [scrims]
      summary: Leave a scrim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Left scrim
  /scrims/{id}/signups:
    get:
      operationId: listScrimSignups
      tags: [scrims]
      summary: List scrim signups
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Signups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScrimSignup"

  # ── LEADERBOARD ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Scrim:
      type: object
      required: [id, opponentName, scheduledAt, gameMode, requiredPlayers, status, createdBy, createdAt]
      properties:
        id:
          type: string
        opponentName:
          type: string
        scheduledAt:
          type: string
        gameMode:
          type: string
        requiredPlayers:
          type: integer
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
        result:
          type: ["string", "null"]
          enum: ["win", "loss", "draw", null]
        createdBy:
          type: string
        createdAt:
          type: string
        signupCount:
          type: integer
    ScrimInput:
      type: object
      required: [opponentName, scheduledAt, gameMode, requiredPlayers]
      properties:
        opponentName:
          type: string
        scheduledAt:
          type: string
        gameMode:
          type: string
        requiredPlayers:
          type: integer
        notes:
          type: string
    ScrimUpdate:
      type: object
      properties:
        opponentName:
          type: string
        scheduledAt:
          type: string
        gameMode:
          type: string
        requiredPlayers:
          type: integer
        notes:
          type: string
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
        result:
          type: string
          enum: [win, loss, draw]
    ScrimSignup:
      type: object
      required: [id, scrimId, memberId, memberName, memberRole, signupStatus, createdAt]
      properties:
        id:
          type: string
        scrimId:
          type: string
        memberId:
          type: string
        memberName:
          type: string
        memberRole:
          type: string
        signupStatus:
          type: string
          enum: [pending, accepted, declined]
        createdAt:
          type: string
...